// src/modules/admin/types.ts
export interface ManagerInput {
  fullName: string;
  email: string;
  phone: string;
  position: string;
}

export interface CustomerInput {
  name: string;
  contactPerson: string;
  email: string;
  inn: string;
  phone: string;
  address: string;
  comment: string;
}

export interface EmployeeInput {
  fullName: string;
  email: string;
  phone: string;
  position: string;
}